charset_for_atlus_script_tool/requirements.txt
annoy==1.17.1
fontTools==4.38.0
matplotlib==3.4.3
numpy==1.23.5
opencv_python==4.5.5.62
Pillow==9.3.0
